RF: Move get_cmd_summaries from cmdline/main to interface/base
This will be used outside of the command-line interface.
