+++ datalad/interface/base.py
@@ -93,6 +93,24 @@ def get_interface_groups(include_plugins=False):
     return grps
 
 
+def get_cmd_summaries(grp_short_descriptions, interface_groups, width=79):
+    cmd_summary = []
+    for grp in sorted(interface_groups, key=lambda x: x[1]):
+        grp_descr = grp[1]
+        grp_cmds = grp_short_descriptions[grp[0]]
+
+        cmd_summary.append('\n*%s*\n' % (grp_descr,))
+        for cd in grp_cmds:
+            cmd_summary.append('  %s\n%s'
+                               % ((cd[0],
+                                   textwrap.fill(
+                                       cd[1].rstrip(' .'),
+                                       width - 5,
+                                       initial_indent=' ' * 6,
+                                       subsequent_indent=' ' * 6))))
+    return cmd_summary
+
+
 def dedent_docstring(text):
     """Remove uniform indentation from a multiline docstring"""
     # Problem is that first line might often have no offset, so might
